ParcelDraft to interface. #942
