+++ sphere-models/src/it/java/io/sphere/sdk/orders/commands/OrderUpdateCommandTest.java
@@ -18,7 +18,6 @@ import io.sphere.sdk.states.State;
 import io.sphere.sdk.states.StateType;
 import io.sphere.sdk.test.IntegrationTest;
 import io.sphere.sdk.test.SphereTestUtils;
-import io.sphere.sdk.utils.SphereInternalUtils;
 import org.assertj.core.api.Assertions;
 import org.assertj.core.api.Condition;
 import org.junit.Test;
