+++ sphere-models/src/main/java/io/sphere/sdk/orders/ParcelDraft.java
@@ -1,39 +1,23 @@
 package io.sphere.sdk.orders;
 
-import io.sphere.sdk.models.Base;
-
 import javax.annotation.Nullable;
 
-public class ParcelDraft extends Base {
+public interface ParcelDraft {
     @Nullable
-    private final ParcelMeasurements measurements;
-    @Nullable
-    private final TrackingData trackingData;
-
-    private ParcelDraft(@Nullable final ParcelMeasurements measurements, @Nullable final TrackingData trackingData) {
-        this.measurements = measurements;
-        this.trackingData = trackingData;
-    }
-
-    public static ParcelDraft of(final ParcelMeasurements measurements, final TrackingData trackingData) {
-        return new ParcelDraft(measurements, trackingData);
-    }
-
-    public static ParcelDraft of(final ParcelMeasurements measurements) {
-        return new ParcelDraft(measurements, null);
-    }
-
-    public static ParcelDraft of(final TrackingData trackingData) {
-        return new ParcelDraft(null, trackingData);
-    }
+    ParcelMeasurements getMeasurements();
 
     @Nullable
-    public ParcelMeasurements getMeasurements() {
-        return measurements;
+    TrackingData getTrackingData();
+
+    static ParcelDraft of(final ParcelMeasurements measurements, final TrackingData trackingData) {
+        return new ParcelDraftImpl(measurements, trackingData);
     }
 
-    @Nullable
-    public TrackingData getTrackingData() {
-        return trackingData;
+    static ParcelDraft of(final ParcelMeasurements measurements) {
+        return new ParcelDraftImpl(measurements, null);
+    }
+
+    static ParcelDraft of(final TrackingData trackingData) {
+        return new ParcelDraftImpl(null, trackingData);
     }
 }
